oh yeah²

+++ LibraryAPI/Services/PublisherService.cs
@@ -4,7 +4,6 @@ using LibraryAPI.Dtos.Publishers;
 using LibraryAPI.Dtos.Validations;
 using LibraryAPI.Models;
 using LibraryAPI.Services.Interface;
-using Microsoft.AspNetCore.Mvc;
 
 namespace LibraryAPI.Services
 {
@@ -12,40 +11,41 @@ namespace LibraryAPI.Services
     {
         private readonly IPublisherRepository _publisherRepository;
         private readonly IMapper _mapper;
+        private readonly IBookRepository _bookRepository;
 
-        public PublisherService(IPublisherRepository publisherRepository, IMapper mapper)
+        public PublisherService(IPublisherRepository publisherRepository, IMapper mapper, IBookRepository bookRepository)
         {
             _publisherRepository = publisherRepository;
             _mapper = mapper;
+            _bookRepository = bookRepository;
         }
 
         public async Task<ResultService> CreateAsync(CreatePublisherDto createPublisherDto)
         {
-            var result = new PulisherDtoValidator().Validate(createPublisherDto);
-            if (!result.IsValid)
-                return ResultService.RequestError<CreatePublisherDto>("Problemas da validade!: ", result);
+            var validation = new PulisherDtoValidator().Validate(createPublisherDto);
+            if (!validation.IsValid)
+                return ResultService.RequestError<CreatePublisherDto>("Problemas da validade!: ", validation);
 
-            var samePublisher = await _publisherRepository.GetPublisherByName(createPublisherDto.Name);
-
-            if (samePublisher != null)
+            var sameName = await _publisherRepository.GetPublisherByName(createPublisherDto.Name);
+            if (sameName != null)
                 return ResultService.Fail<CreatePublisherDto>("Editora já existente.");
 
             var publisher = _mapper.Map<Publisher>(createPublisherDto);
             await _publisherRepository.Add(publisher);
             return ResultService.Ok("Editora cadastrada.");
-            
+
         }
 
         public async Task<ResultService<ICollection<PublisherDto>>> GetAsync()
         {
             var publisher = await _publisherRepository.GetAllPublishers();
-            return ResultService.Ok<ICollection<PublisherDto>>(_mapper.Map<ICollection<PublisherDto>>(publisher));
+            return ResultService.Ok(_mapper.Map<ICollection<PublisherDto>>(publisher));
         }
 
         public async Task<ResultService<PublisherDto>> GetByIdAsync(int id)
         {
             var publisher = await _publisherRepository.GetPublisherById(id);
-            if(publisher == null)
+            if (publisher == null)
             {
                 return ResultService.Fail<PublisherDto>("Editora não encontrada!");
             }
@@ -54,19 +54,21 @@ namespace LibraryAPI.Services
 
         public async Task<ResultService> UpdateAsync(PublisherDto publisherDto)
         {
-            if (publisherDto == null)
+            var publisher = await _publisherRepository.GetPublisherById(publisherDto.Id);
+            if (publisher == null)
                 return ResultService.Fail<PublisherDto>("Editora não encontrada.");
 
             var validation = new UpdatePulisherDtoValidator().Validate(publisherDto);
             if (!validation.IsValid)
                 return ResultService.RequestError(validation);
 
-            var publisher = await _publisherRepository.GetPublisherById(publisherDto.Id);
+            var sameName = await _publisherRepository.GetPublisherByName(publisherDto.Name);
+            if (sameName != null && sameName.Id != publisher.Id)
+                return ResultService.Fail<CreatePublisherDto>("Editora já existente.");
 
-            publisher = _mapper.Map<PublisherDto, Publisher>(publisherDto, publisher);
+            publisher = _mapper.Map(publisherDto, publisher);
             await _publisherRepository.Update(publisher);
             return ResultService.Ok("Editora atualizada");
-
         }
 
         public async Task<ResultService> DeleteAsync(int id)
@@ -75,9 +77,12 @@ namespace LibraryAPI.Services
             if (publisher == null)
                 return ResultService.Fail<PublisherDto>("Editora não encontrado.");
 
-            await _publisherRepository.Delete(publisher);
-            return ResultService.Ok("A editora foi deletada");
-        }
+            var bookAssociation = await _bookRepository.GetPublisherAssociate(id);
+            if (bookAssociation.Count > 0)
+                return ResultService.Fail<PublisherDto>("Editora está associada a livros.");
 
+            await _publisherRepository.Delete(publisher);
+            return ResultService.Ok("Editora foi deletada");
+        }
     }
 }
